reorganizing JS files
